requirements.txt
faiss_cpu==1.10.0
fastapi==0.115.12
matplotlib==3.10.1
numpy==2.2.5
Pillow==11.2.1
Requests==2.32.3
scikit_learn==1.6.1
streamlit==1.43.0
torch==2.6.0
torchvision==0.21.0
transformers==4.49.0
